requirements.txt
earthengine-api==0.1.378
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-cloud-storage==2.13.0
oauth2client==4.1.3
pandas==2.1.4
numpy==1.24.3
geopandas==0.14.0
rasterio==1.3.9
matplotlib==3.8.3
seaborn==0.13.1
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1
